Significant Progress (See Details)
-Adding and deleting of charter trips now functional
-When assigning employees to new charter trips, UI gives feedback to user depending if the user already assigned enough employees to meet the requirements of the properties of the chosen aircraft. UI also alerts user of extra crew members.
-Further UI improvements
-Adding and deleting flights to charter trips now functional
-Charter trips "Total" properties as well as their Origin and Final Destination now updates depending on the flight legs assigned to them.

+++ RC Charter2/RC Charter2/Models/CharterTrip.cs
@@ -21,7 +21,7 @@ namespace RC_Charter2.Models
 		public int? AircraftNumber { get; set; }
 		public int? CustomerId { get; set; }
 		public float? TotalDistanceFlown { get; set; }
-		public int? TotalWaitingTime { get; set; }
+		public float? TotalWaitingTime { get; set; }
 		public float? TotalFuelUsage { get; set; }
 		public string Origin { get; set; }
 		public string FinalDestination { get; set; }
